generate_helm_release: Fix cached repo dir
If a release manager had previously tagged a release, then they would
have already checkout out a copy of the target project into the local
directory. In this scenario however the checkout command attempted to
create a new branch with the name of the new release, but the content of
the branch would be the currently checked out SHA.

This branch requires that the upstream project has already tagged the
target release and that this repo is just pulling that target version in
order to generate the Helm release.

Fixes: a9d85774177e ("generate_helm_release: use -B for checkout")
Signed-off-by: Joe Stringer <joe@cilium.io>