gen: Fix invalid git references
These lines were intended to check out the latest main branch of the
charts repo, but instead it was trying to find the remote for the target
project to release (cilium or tetragon). Furthermore the target variable
was intended to be HEAD, but it was undefined here. Fix both issues.

Solves an issue where the script would fail with the following error:

    fatal: no path specified; see 'git help pull' for valid url syntax

Fixes: 77d0f357d783 ("gen: Check out and sync the local branch first")
Reported-by: Tim Horner <timothy.horner@isovalent.com>
Signed-off-by: Joe Stringer <joe@cilium.io>